requirements.txt
# -----------------------------------------------------------------------
# Core KFP
# -----------------------------------------------------------------------
kfp>=2.5.0
kfp-pipeline-spec>=0.3.0

# -----------------------------------------------------------------------
# Data processing
# -----------------------------------------------------------------------
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0
scipy>=1.11.0

# -----------------------------------------------------------------------
# ML frameworks (install only what you need)
# -----------------------------------------------------------------------
scikit-learn>=1.3.0
xgboost>=2.0.0
# tensorflow>=2.15.0      # Uncomment for TF/Keras
# torch>=2.1.0            # Uncomment for PyTorch
# torchvision>=0.16.0     # Uncomment for PyTorch vision
# keras>=3.0.0            # Uncomment for standalone Keras
# flaml>=2.1.0            # Uncomment for AutoML (FLAML)
# auto-sklearn>=0.15.0    # Uncomment for AutoML (auto-sklearn)

# -----------------------------------------------------------------------
# Hyperparameter tuning
# -----------------------------------------------------------------------
optuna>=3.4.0

# -----------------------------------------------------------------------
# Serialization / config
# -----------------------------------------------------------------------
joblib>=1.3.0
pyyaml>=6.0.1

# -----------------------------------------------------------------------
# Deployment / cloud (optional)
# -----------------------------------------------------------------------
# google-cloud-aiplatform>=1.38.0
# google-cloud-build>=3.22.0
# google-cloud-storage>=2.13.0
# google-cloud-bigquery>=3.13.0
# google-cloud-artifactregistry>=1.9.0

# -----------------------------------------------------------------------
# Visualization
# -----------------------------------------------------------------------
matplotlib>=3.8.0

# -----------------------------------------------------------------------
# Networking / notifications
# -----------------------------------------------------------------------
requests>=2.31.0
